mruby: §E row 7 — fuzz the data-driven parsers (libFuzzer, ASan+UBSan)
Add five libFuzzer harnesses (testing/mruby/fuzz/) over the parsers most
exposed to malformed input: the love.data pack/unpack format engine, the
love.filesystem require-path resolver, and the image/sound/font decoders.
One source compiled per target (-DFZ_<T>), sharing one mruby VM and driving
each target through the real Ruby binding; built via `make fuzz` (clang).

Fix found: signed-int overflow miscomputing a decoded image's byte size —
`width * height * 4` done in `int` and only then widened to the size_t size,
so a crafted image wraps the recorded size while the real buffer differs (a
heap-overflow primitive reachable from Love::Image.new_image_data). Same
pattern in four shared magpie sites; widened the multiply to size_t in
STBHandler (8-bit + HDR), EXRHandler (the `new T[]` allocation size),
ImageData::getSize, and PNGHandler. Upstream/shared code, so the Lua build
benefits too. Suite still 24/24; the UBSan reproducer now decodes cleanly.

Also documented (not fixed — upstream policy, present in mainline): two
single-input allocation-amplification DoS surfaces (stb TGA sizing from
header dimensions; pack's cN field), with reproducers under fuzz/findings/.

+++ src/modules/image/ImageData.cpp
@@ -240,7 +240,7 @@ love::filesystem::FileData *ImageData::encode(FormatHandler::EncodedFormat encod
 
 size_t ImageData::getSize() const
 {
-	return size_t(getWidth() * getHeight()) * getPixelSize();
+	return size_t(getWidth()) * size_t(getHeight()) * getPixelSize();
 }
 
 void *ImageData::getData() const
